Refactor: split MainActivity.kt and DashboardScreen.kt into focused files
- ShellUtils.kt: constants, suExec, suExecAsync, ensureSelfInTargets,
  cleanupStaleZygiskStatus
- AppPickerScreen.kt: AppEntry, InstalledModules, AppPickerScreen, AppRow,
  LayerChip, AppsHelpDialog, buildSaveCommand, buildUidResolver
- DashboardData.kt: domain types (ModuleState, LsposedState, ProtectionCheck,
  etc.), loadDashboardState, protection check functions
- MainActivity.kt: only Activity, theme, MainScreen navigation, root screens
- DashboardScreen.kt: only UI composables

No behavioral changes. MainActivity 981→343 lines, DashboardScreen 1363→623.
